Implemented variable size calculation

+++ examples/fibonacci.c
@@ -1,11 +1,11 @@
 // Returns the 10th number of the fibonacci sequence
 int main() {
-  int f0 = 0;
-  int f1 = 1;
+  long f0 = 0;
+  long f1 = 1;
 
-  int i = 2;
+  long i = 2;
   while (1) {
-    int tmp = f0 + f1;
+    long tmp = f0 + f1;
     f0 = f1;
     f1 = tmp;
     ++i;
